python/requirement.txt
# Python project dependencies
cryptography
Pillow